feat(Autenticação): finalizando a autenticação de usuário

+++ controllers/usuarioController.js
@@ -23,6 +23,12 @@ const UsuarioController = () =>{
 
     const readAllUsers = async (req, res) => {
         try{
+            const userId = req.user.id;
+
+            if(req.user.tipo !== "admin"){
+                return res.status(403).json({error: "Acesso não concedido."});
+            }
+
             let usuariosInfo = await Usuario.findAll({});
             res.status(200).send(usuariosInfo);
         } catch(err){
@@ -34,6 +40,11 @@ const UsuarioController = () =>{
     const readById = async (req, res) => {
         try{
             let id = req.params.id;
+
+            if(req.user.tipo !== "admin" && req.user.id !== parseInt(id)) {
+                return res.status(403).json({error: "Acesso não concedido."});
+            }
+
             let usuarioInfo = await Usuario.findOne({where: {idUsuario: id}});
             res.status(200).send(usuarioInfo);
         } catch(err){
@@ -45,6 +56,11 @@ const UsuarioController = () =>{
     const updateUser = async (req, res) => {
         try{
             let id = req.params.id;
+
+            if(req.user.tipo !== "admin" && req.user.id !== parseInt(id)) {
+                return res.status(403).json({error: "Acesso não concedido."});
+            }
+
             const usuario = await Usuario.update(req.body, {where: {idUsuario: id}});
             res.status(200).send(usuario);
         } catch(err){
@@ -56,6 +72,11 @@ const UsuarioController = () =>{
     const deleteUser = async (req, res) => {
         try{
             let id = req.params.id;
+
+            if(req.user.tipo !== "admin" && req.user.id !== parseInt(id)) {
+                return res.status(403).json({error: "Acesso não concedido."});
+            }
+            
             const usuario = await Usuario.destroy({where: {idUsuario: id}});
             res.status(200).send('User deleted succesfully!');
         } catch(err){
